+++ CHANGELOG.md
@@ -7,9 +7,43 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-## [Unreleased](https://github.com/global-121/121-platform/compare/v1.29.3...master)
+## [Unreleased](https://github.com/global-121/121-platform/compare/v1.30.4...master)
+
+### Added
+
+- 2021-12-23: Conditional Maintenance-mode with status-check for deploy-script  
+  Make sure to define `GLOBAL_121_STATUS_URL` in `./tools/.env` when using a custom hostname/port-number for the 121-service
+- 2021-12-22: Add configuration flags for `webhook.js`-script:  
+  `DEPLOY_PRE_RELEASE` and `DEPLOY_RELEASE` to enable auto-deployment of these type of (pre-)release payloads.
+- 2021-12-28: Added 'manual deployment' via the `webhook.js`-script. See [tools/README](./tools/README.md#webhook-script).
+- 2021-12-29: Added `whatsappGenericMessage` to the NL program notifications.
+
+## Changed
+
+- 2021-12-28: "Automatic hotfix deployment" is now an opt-in behavior by setting the ENV-variable `DEPLOY_PATCH=1`. See [tools/README](./tools/README.md#webhook-script).
 
 ---
+
+## [1.30.4](https://github.com/global-121/121-platform/compare/v1.30.3...v1.30.4) - 2021-12-22
+
+- Test release
+
+## [1.30.3](https://github.com/global-121/121-platform/compare/v1.30.2...v1.30.3) - 2021-12-22
+
+- Test release
+
+## [1.30.2](https://github.com/global-121/121-platform/compare/v1.30.1...v1.30.2) - 2021-12-22
+
+### Added
+
+- 2021-12-22: Added pre-release option to webhook-script
+
+## [1.30.1](https://github.com/global-121/121-platform/compare/v1.30.0...v1.30.1) - 2021-12-22
+
+- Test release
+
+## [1.30.0](https://github.com/global-121/121-platform/compare/v1.29.3...v1.30.0) - 2021-12-22
+
 ## [1.29.3](https://github.com/global-121/121-platform/compare/v1.29.2...v1.29.3) - 2021-12-17
 
 ## [1.29.2](https://github.com/global-121/121-platform/compare/v1.29.1...v1.29.2) - 2021-12-15